3/5/2018 Death Animations!
Enemies can die, they fall over and can no longer interfere with pathfinding. Not quite a "As an entity, I want to be able to show that I have died if I have 0 HP", but it can easily be converted into that in the future (currently tied to enemyBasicStats, which should be converted in the future to be more universal.