The Flickering Continues (read the description)
I've figured out why the UI for the inventory flickers, it's to do with PGI being sort-of out of date with Unity doing serious GUI changes across the period of time it was developed.

Should look at updating to instead use this version of PGI instead, which may solve the problem:
https://github.com/Slugronaut/PGI

On the flipside, if the code is cross-compatible as it should be and this version no longer has the flickering, I think it could only require 10 to 20 minutes to make sure it all works.

The good news is we can now correctly drop items anyway. I was able to get up just prior to the Equipment section of the Getting Started document.