Improvements
Funnily enough, Jaime wrote back and stated to not worry too much about the flickering inventory ui. On that basis, we now have Equipment Slots, and a way to equip swords. The prefab for heroes will need a bit of extensive updating, and the ui manager will need to be updated with the slot script from the other branch in order to be able to correctly assign inventories to each hero, but on the flipside equipping weapons and potions does in fact work.